refactor : add constructor take File
