+++ src/main/java/nextstep/courses/domain/session/CoverImage.java
@@ -1,0 +1,94 @@
+package nextstep.courses.domain.session;
+
+import java.awt.image.BufferedImage;
+import java.io.File;
+import java.io.IOException;
+import java.util.Arrays;
+import java.util.HashSet;
+import java.util.Objects;
+import java.util.Set;
+
+import javax.imageio.ImageIO;
+
+public class CoverImage {
+
+    private static final Set<String> ALLOWED_FILE_EXTENSIONS
+            = new HashSet<>(Arrays.asList("gif", "jpeg", "jpg", "png", "svg"));
+    private static final int MAX_FILE_SIZE = 1024 * 1024;
+
+    private static final int MIN_WIDTH = 300;
+    private static final int MIN_HEIGHT = 200;
+
+    private final File source;
+
+    private CoverImage(File imagePath) {
+        this.source = imagePath;
+    }
+
+    public static CoverImage of(String imagePath) {
+        return of(new File(imagePath));
+    }
+
+    public static CoverImage of(File imageFile) {
+        validate(imageFile);
+        return new CoverImage(imageFile);
+    }
+
+    private static void validate(File imageFile) {
+        validateFile(imageFile);
+        validateImagePixel(imageFile);
+    }
+
+    private static void validateFile(File imageFile) {
+        if (!ALLOWED_FILE_EXTENSIONS.contains(parseExtension(imageFile.getName()))) {
+            throw new IllegalArgumentException("invalid");
+        }
+        if (imageFile.length() > MAX_FILE_SIZE) {
+            throw new IllegalArgumentException("invalid");
+        }
+    }
+
+    private static String parseExtension(String filename) {
+        int index = filename.lastIndexOf(".");
+        return (index > -1) ? filename.substring(index + 1).toLowerCase() : "";
+    }
+
+    public static void validateImagePixel(File file) {
+        try {
+            BufferedImage bufferedImage = ImageIO.read(file);
+            int width = bufferedImage.getWidth();
+            int height = bufferedImage.getHeight();
+
+            if (width < MIN_WIDTH || height < MIN_HEIGHT) {
+                throw new IllegalArgumentException("강의 커버 이미지는 가로 300px 이상, 세로 200px 이상이여야 합니다.");
+            }
+
+            if (getRadio(width, height) != getRadio(3, 2)) {
+                throw new IllegalArgumentException("강의 커버 이미지는 가로 세로 비율이 3:2 여야 합니다.");
+            }
+        } catch (IOException e) {
+            throw new IllegalStateException(e);
+        }
+    }
+
+    private static double getRadio(int width, int height) {
+        return (double) width / height;
+    }
+
+    @Override
+    public boolean equals(Object o) {
+        if (this == o) {
+            return true;
+        }
+        if (o == null || getClass() != o.getClass()) {
+            return false;
+        }
+        CoverImage that = (CoverImage) o;
+        return Objects.equals(source, that.source);
+    }
+
+    @Override
+    public int hashCode() {
+        return Objects.hashCode(source);
+    }
+}
